+++ src/Formatter/SlackFormatterInterface.php
@@ -2,11 +2,13 @@
 
 namespace Webthink\MonologSlack\Formatter;
 
+use Monolog\Formatter\FormatterInterface;
+
 /**
  * This is an Interface that all formatters must extend in order to be passed into SlackWebhookHandler
  *
  * @author George Mponos <gmponos@gmail.com>
  */
-interface SlackFormatterInterface
+interface SlackFormatterInterface extends FormatterInterface
 {
 }