+++ src/Formatter/SlackLineFormatter.php
@@ -21,13 +21,13 @@ class SlackLineFormatter extends LineFormatter implements SlackFormatterInterfac
     /**
      * User icon e.g. 'ghost', 'http://example.com/user.png'
      *
-     * @var string
+     * @var string|null
      */
     protected $emoji;
 
     /**
      * @param string|null $username
-     * @param string $emoji
+     * @param string|null $emoji
      */
     public function __construct(string $username = null, string $emoji = null)
     {
@@ -40,7 +40,7 @@ class SlackLineFormatter extends LineFormatter implements SlackFormatterInterfac
      * @param array $record
      * @return array
      */
-    public function format(array $record)
+    public function format(array $record): array
     {
         $data['text'] = parent::format($record);