+++ src/Formatter/SlackLongAttachmentFormatter.php
@@ -13,9 +13,9 @@ class SlackLongAttachmentFormatter extends AbstractSlackAttachmentFormatter
 {
     /**
      * @param array $record
-     * @return mixed
+     * @return array
      */
-    protected function formatFields(array $record)
+    protected function formatFields(array $record): array
     {
         $result = [];
         foreach ($record as $key => $value) {